# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
}

export interface AuthResult {
  user: User;
}

export interface DashboardStats {
  totalProducts: number;
  inStock: number;
  lowStock: number;
  outOfStock: number;
  inventoryValue: number;
}

export interface CategoryDistribution {
  category: string;
  count: number;
  percentage: number;
}

export interface MonthlyTrend {
  month: string;
  restocked: number;
  lowStock: number;
}

export interface TopCategory {
  category: string;
  value: number;
  count: number;
}

export interface Product {
  id: number;
  name: string;
  categoryId: number;
  categoryName: string;
  quantity: number;
  unit: string;
  price: number;
  lowStockThreshold?: number;
  status: string;
  /** @nullable */
  barcode: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  brand?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ProductInput {
  name: string;
  categoryId: number;
  quantity: number;
  unit: string;
  price: number;
  lowStockThreshold?: number;
  barcode?: string;
  description?: string;
  brand?: string;
}

export interface ProductUpdate {
  name?: string;
  categoryId?: number;
  quantity?: number;
  unit?: string;
  price?: number;
  lowStockThreshold?: number;
  barcode?: string;
  description?: string;
  brand?: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  productCount: number;
}

export interface CategoryInput {
  name: string;
  description?: string;
}

export interface CategoryUpdate {
  name?: string;
  description?: string;
}

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: string;
  quantity: number;
  /** @nullable */
  reason: string | null;
  createdAt: string;
}

export interface StockMovementInput {
  productId: number;
  type: string;
  quantity: number;
  reason?: string;
}

export interface StockAlert {
  id: number;
  productId: number;
  productName: string;
  categoryName: string;
  quantity: number;
  lowStockThreshold: number;
  status: string;
  updatedAt: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  status: string;
  totalAmount: number;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerId: number;
  items: OrderItemInput[];
}

export interface OrderUpdate {
  status?: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  totalOrders?: number;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  phone: string;
  email?: string;
  address?: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  address?: string;
}

export type ListProductsParams = {
category?: string;
status?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListStockMovementsParams = {
productId?: number;
type?: string;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
status?: string;
page?: number;
limit?: number;
};

export type ListCustomersParams = {
search?: string;
page?: number;
limit?: number;
};