# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerId: number;
  customerName: string;
  status: string;
  totalAmount: number;
  items: OrderItem[];
  createdAt: string;
}